+++ ThorgApp/Interfaces/IProviderConfig.cs
@@ -25,7 +25,7 @@ namespace GolemUI.Interfaces
         public void UpdateActiveCpuThreadsCount(int threadsCount);
         void UpdateWalletAddress(string? walletAddress = null);
         void UpdateNodeName(string? value);
-        void UpdateCharity(string? wallet, float? procentage);
+        void UpdateCharity(string? wallet, double? procentage);
 
         Task Prepare(bool isGpuCapable, bool isLowMemoryMode);
     }
